refactor(auth): share metadata parse helpers across packages
