+++ sdk/cliproxy/auth/types.go
@@ -4,15 +4,14 @@ import (
 	"context"
 	"crypto/sha256"
 	"encoding/hex"
-	"encoding/json"
 	"net/http"
 	"net/url"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
 
 	baseauth "github.com/router-for-me/CLIProxyAPI/v6/internal/auth"
+	"github.com/router-for-me/CLIProxyAPI/v6/internal/util"
 )
 
 // PostAuthHook defines a function that is called after an Auth record is created
@@ -328,61 +327,11 @@ func (a *Auth) RequestRetryOverride() (int, bool) {
 }
 
 func parseBoolAny(val any) (bool, bool) {
-	switch typed := val.(type) {
-	case bool:
-		return typed, true
-	case string:
-		trimmed := strings.TrimSpace(typed)
-		if trimmed == "" {
-			return false, false
-		}
-		parsed, err := strconv.ParseBool(trimmed)
-		if err != nil {
-			return false, false
-		}
-		return parsed, true
-	case float64:
-		return typed != 0, true
-	case json.Number:
-		parsed, err := typed.Int64()
-		if err != nil {
-			return false, false
-		}
-		return parsed != 0, true
-	default:
-		return false, false
-	}
+	return util.ParseBoolAny(val)
 }
 
 func parseIntAny(val any) (int, bool) {
-	switch typed := val.(type) {
-	case int:
-		return typed, true
-	case int32:
-		return int(typed), true
-	case int64:
-		return int(typed), true
-	case float64:
-		return int(typed), true
-	case json.Number:
-		parsed, err := typed.Int64()
-		if err != nil {
-			return 0, false
-		}
-		return int(parsed), true
-	case string:
-		trimmed := strings.TrimSpace(typed)
-		if trimmed == "" {
-			return 0, false
-		}
-		parsed, err := strconv.Atoi(trimmed)
-		if err != nil {
-			return 0, false
-		}
-		return parsed, true
-	default:
-		return 0, false
-	}
+	return util.ParseIntAny(val)
 }
 
 func (a *Auth) AccountInfo() (string, string) {
@@ -520,49 +469,5 @@ func ProviderRefreshLead(provider string, runtime any) *time.Duration {
 }
 
 func parseTimeValue(v any) (time.Time, bool) {
-	switch value := v.(type) {
-	case string:
-		s := strings.TrimSpace(value)
-		if s == "" {
-			return time.Time{}, false
-		}
-		layouts := []string{
-			time.RFC3339,
-			time.RFC3339Nano,
-			"2006-01-02 15:04:05",
-			"2006-01-02 15:04",
-			"2006-01-02T15:04:05Z07:00",
-		}
-		for _, layout := range layouts {
-			if ts, err := time.Parse(layout, s); err == nil {
-				return ts, true
-			}
-		}
-		if unix, err := strconv.ParseInt(s, 10, 64); err == nil {
-			return normaliseUnix(unix), true
-		}
-	case float64:
-		return normaliseUnix(int64(value)), true
-	case int64:
-		return normaliseUnix(value), true
-	case json.Number:
-		if i, err := value.Int64(); err == nil {
-			return normaliseUnix(i), true
-		}
-		if f, err := value.Float64(); err == nil {
-			return normaliseUnix(int64(f)), true
-		}
-	}
-	return time.Time{}, false
-}
-
-func normaliseUnix(raw int64) time.Time {
-	if raw <= 0 {
-		return time.Time{}
-	}
-	// Heuristic: treat values with millisecond precision (>1e12) accordingly.
-	if raw > 1_000_000_000_000 {
-		return time.UnixMilli(raw)
-	}
-	return time.Unix(raw, 0)
+	return util.ParseTimeAny(v)
 }
